+++ MaintenanceManager/CHANGELOG.md
@@ -16,6 +16,18 @@ All notable changes to this RDK Service will be documented in this file.
 
 * For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
 
+## [1.0.43] - 2025-005-05
+### Fixed
+- Fixed stopMaintenance() jsonRPC API to set MAINTENANCE_ERROR only for an actively running Maintenance.
+
+## [1.0.42] - 2025-04-15
+### Added
+- Added streamlined logging to journal from MaintenanceManager using Systemd journal API.
+
+## [1.0.41] - 2025-03-06
+### Added
+- Added reliable Maintenance Task Status handling and reporting
+
 ## [1.0.40] - 2025-02-11
 ### Removed
 - Changed Maintenance Manager Task execution thread to honour WhoAmI only in Unsolicited Maintenance.